Implement testing
Let's enjoy what we're doing... Frankly the current state is disgusting!
:troll:

I've decided to go with BDD as all of us are familiar with this either
from `ruby` or `go`. `mocha` and `chi` are pretty good for that and are
supported by already existing `hubot-test-helper`.

I have let the `.eslintrc.json` file know, that we're using `mocha`, so
we we can keep on going with our chosen standards...