Catchup with the tests
We'd like to go forward with the confidence of application keep on
working. Ideal situation would be to hold a set of tests, covering each
line of code in our system. This however proves to be difficult for
different reasons.

For instance, I'd like to test that `lit` command when successful would
create a message in a separate channel. Sounds simple, yet the hubot
testing helper does not support the functionality of multiple rooms.
This issue has already been raised with upstream [1] and needs yet to be
resolved.

In order to keep hubot happy, I had to separate the script tests into a
separate directory. In fact, two of them, as I've separated the response
messages from script file, to be DRY about it when using in tests. The
`.messages.js` file is obsolete and _should_ be removed once we come out
with ideal configuration system.

I took the lousy approach of implementing tests for the other functions.
I'm sorry.

[1] https://github.com/mtsmfm/hubot-test-helper/issues/32