Implement robotk review command
We'd like to receive some data from an API regarding a game we're currently
talking about in the chat. Robotk could provide us with that information if we
ask nicely.

`nock` is a testing library to mock any http request. AFAIK it doesn't
make the request, but it captures any directed at this direction, then
composes own response later on.